requirements-dev.txt
pep8
pyflakes
coverage
pytest<3.0,>=2.9.0
pytest-xdist
pytest-cov
requests_testadapter
fake-go-contacts>=0.1.9
pytest-django<3.0,>=2.9.0
flake8
